fix(ax-abi-windows): boot CPython 3.14 python.exe under the Windows personality
本轮把 Windows 人格从"python.exe 起不来"推进到"CPython 3.14.7 完整启动、执行 -c、exit 0"，并顺带确认 macOS smoke 与 Linux 现状。阶段性成果，先提交存档。

## Windows

连破真 ABI 缺陷：

- `CompareStringOrdinal` 未实现（getpath 比对/规范化路径要用），补 `locale::compare_string_ordinal`（镜像 CompareStringW，含 ignore-case 折叠）。
- `create_file` 把 Win32 DWORD `dwCreationDisposition`/`dwDesiredAccess`/`dwFlagsAndAttributes` 当整个 usize 处理；x64 栈槽高 32 位可能带垃圾（如 `0x1_0000_0003`），使 getpath 的 isfile/os.stat 误命中 `ERROR_INVALID_PARAMETER`→EINVAL 而崩。修为参数绑定截断到 `u32`。这是 getpath EINVAL 的真根因。
- `CreateEventA/W` 未实现（signal init 只需有效非空句柄），补最小实现。
- win-py 换用 CPython 3.14.7 embeddable：补 `SleepEx` 及 3.14 新增的 10 个 kernel32 导入（CreateFileMappingA、CreateToolhelp32Snapshot、FlushInstructionCache、GetCurrentThreadStackLimits、Module32FirstW/NextW、OpenThread、ReadProcessMemory、SetThreadStackGuarantee、WriteProcessMemory）。
- prebuild：`pyvenv.cfg`→`python._pth`、`python313`→`python314`、并把扩展 python-lang 套件装到镜像 `/suite`（qemu toml 暂留 `WIN-PY-OK` 里程碑态）。

到"跑完整扩展 py 套件"还差一整批子系统：`VirtualFree`（现未实现，import 写 .pyc 即 MemoryError）、`CreateProcessW`/subprocess（run_all.py 每模块起子解释器）、线程、ws2_32 套接字、multiprocessing。

## macOS

smoke 通过：`mac-abi` 手工 Mach-O（直接 Darwin write+exit 系统调用，无 dyld/libSystem）经 `ax-abi-darwin` → `MAC-ABI-OK`、exit 0。到扩展套件还差 dyld + chained fixups 与 Mach traps（mach_task_self/mach_vm_allocate/mach_msg）。

## Linux

早前已亲验通过补充后的 CPython 3.14 套件 23/23（x86_64）。解耦部分完成：`ax-abi-linux` 已有 `elf.rs`、syscall 分发已组件化、能力注册用 `linkme`；内核 `mm/loader.rs` 仍在，尚未彻底剥离（进行中，目标是内核无关、Linux 仅作人格实现示例）。

+++ components/ax-abi-windows/src/win32/file.rs
@@ -152,8 +152,17 @@ fn descriptor(handle: usize) -> Result<i32, Ntstatus> {
 /// CreateFileW(lpFileName, dwDesiredAccess, dwShareMode, lpSecurityAttributes,
 /// dwCreationDisposition, dwFlagsAndAttributes, hTemplateFile).
 pub fn create_file(c: &mut Call<'_>) -> Dispatch {
-    let (name, access, sa, creation, attributes) =
-        (c.arg(0), c.arg(1), c.arg(3), c.arg(4), c.arg(5));
+    // CreateFileW's dwDesiredAccess / dwCreationDisposition /
+    // dwFlagsAndAttributes are 32-bit DWORDs; the x64 stack slots carrying
+    // them may leave the upper 32 bits dirty, so truncate before use (a
+    // dirty disposition like 0x1_0000_0003 must read as OPEN_EXISTING).
+    let (name, access, sa, creation, attributes) = (
+        c.arg(0),
+        c.arg(1) as u32 as usize,
+        c.arg(3),
+        c.arg(4) as u32 as usize,
+        c.arg(5) as u32 as usize,
+    );
     let Some(name) = name_at(c, name).filter(|n| !n.is_empty()) else {
         return c.fail(ERROR_PATH_NOT_FOUND, INVALID_HANDLE_VALUE);
     };
